% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-lab-counts.R
\name{load_lab_counts}
\alias{load_lab_counts}
\title{Load enrolment numbers in FoS lab courses}
\usage{
load_lab_counts()
}
\value{
A data frame.
}
\description{
Load data frame of total enrolment numbers and numbers of students registered
with Centre for Accessiblity (CfA) in lab courses of the Faculty of Science
(FoS) from September 2024 to August 2025.
}
