% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-lab-requests.R
\name{load_lab_requests}
\alias{load_lab_requests}
\title{Load status of CfA requests in lab courses}
\usage{
load_lab_requests()
}
\value{
A data frame.
}
\description{
Load data frame of Faculty of Science students registered with Centre for
Accessiblity (CfA) and their accomodation request status in lab courses from
September 2024 to August 2025.
}
