% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-students.R
\name{load_students}
\alias{load_students}
\title{Load CfA students}
\usage{
load_students()
}
\value{
A data frame.
}
\description{
Load data frame of Faculty of Science students registered with Centre for
Accessiblity from September 2024 to August 2025.
}
