% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-undergraduate-lab-accessibility-survey.R
\name{load_undergraduate_lab_accessibility_survey}
\alias{load_undergraduate_lab_accessibility_survey}
\title{Load lab accessibility survey of undergraduates in FoS}
\usage{
load_undergraduate_lab_accessibility_survey()
}
\value{
A data frame.
}
\description{
Load data frame of lab accessibility Qualtrics survey data of undergraduates
in the Faculty of Science (FoS) conducted by Skylight in March 2026.
}
